Refactor and abstract JudgingTab
